requirements.txt
langchain-community==0.3.15
openai==1.59.9
langchain-openai==0.3.2
langchain-core==0.3.31
python-dotenv~=1.0.1
langchain~=0.3.15
requests~=2.32.3